+++ src/main/java/me/ionar/salhack/events/packet/PacketEvent.java
@@ -42,4 +42,14 @@ extends MinecraftEvent {
             super(packet);
         }
     }
+    public PacketSendEvent(Packet<?> packet) {
+		this.packet = packet;
+	}
+
+	public Packet<?> getPacket() {
+		return packet;
+	}
+
+	public void setPacket(Packet<?> packet) {
+		this.packet = packet;
 }
